+++ injectable/example/lib/module/register_module.dart
@@ -1,56 +1,44 @@
 import 'dart:async';
 
+import 'package:example/services/abstract_service.dart';
 import 'package:injectable/injectable.dart';
-import 'package:meta/meta.dart';
 
 @module
 abstract class RegisterModule {
-  // @prod
-  // @LazySingleton(as: Repo)
-  // RepoImpl get repo;
-  //
-  // @dev
-  // Future<Repo> getRepo(LazyService service) {
-  //   return Repo.asyncRepo(service);
-  // }
-  //
-  // List get strings => ['strings'];
+  @prod
+  @LazySingleton(as: Repo, dispose: disposeRepo)
+  RepoImpl get repo;
+
+  @dev
+  Future<Repo> getRepo(LazyService service) {
+    return Repo.asyncRepo(service);
+  }
+
+  @Named("StringsList")
+  List get strings => ['strings'];
 }
 
-// abstract class Repo {
-//   @factoryMethod
-//   static Future<RepoImpl> asyncRepo(LazyService service) async {
-//     await Future.delayed(Duration(seconds: 1));
-//     return RepoImpl(service);
-//   }
-//
-//   void dispose();
-// }
-//
-// class RepoImpl extends Repo {
-//   final LazyService service;
-//
-//   RepoImpl(this.service);
-//
-//   @overrid
-//   void dispose() {
-//     // TODO: imxplement dispose
-//   }
-// }
-
-abstract class AbsService {
-  FutureOr dispose();
+FutureOr disposeRepo(Repo instance) {
+  instance.dispose();
 }
 
-@Singleton(as: AbsService, dispose: Disposable.dispose)
-class DisposableService extends AbsService {
-  // @disposeMethod
+abstract class Repo {
+  @factoryMethod
+  static Future<RepoImpl> asyncRepo(LazyService service) async {
+    await Future.delayed(Duration(seconds: 1));
+    return RepoImpl(service);
+  }
+
+  void dispose();
+}
+
+class RepoImpl extends Repo {
+  final LazyService service;
+
+  RepoImpl(this.service);
+
   @override
-  FutureOr dispose() {}
+  void dispose() {
+    print("Disposing RepoImpl");
+  }
 }
-
-class Disposable {
-  static FutureOr dispose(AbsService repo) {}
-}
-
-FutureOr dispose(AbsService repo, String name) {}